Factor out items selector panel
